+++ src/whats-new/contentProvider.ts
@@ -22,8 +22,42 @@ export class SeparatorsContentProvider implements ContentProvider {
 
     public provideChangeLog(): ChangeLogItem[] {
         const changeLog: ChangeLogItem[] = [];
-        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "2.0.0", releaseDate: "November 2020" } });
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "2.1.0", releaseDate: "May 2021" } });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Support <b>Virtual Workspaces</b>",
+                id: 27,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.NEW,
+            detail: {
+                message: "Support <b>Workspace Trust</b>",
+                id: 26,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Do not show welcome message if installed by Settings Sync",
+                id: 29,
+                kind: IssueKind.Issue
+            }
+        });
+        changeLog.push({
+            kind: ChangeLogKind.INTERNAL,
+            detail: {
+                message: "Security Alert: ini",
+                id: 20,
+                kind: IssueKind.PR,
+                kudos: "dependabot"
+            }
+        });
 
+        changeLog.push({ kind: ChangeLogKind.VERSION, detail: { releaseNumber: "2.0.0", releaseDate: "November 2020" } });
         changeLog.push({
             kind: ChangeLogKind.NEW,
             detail: {